requirement.txt
tqdm
tensorboard
fastapi
uvicorn[standard]
python-multipart
torch-ao
pytest